requirements.txt
Flask
Flask-JWT-Extended
Flask-Cors
Flask-SQLAlchemy
marshmallow
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
